Plot frequency distributions of key shuttle-box metrics

This function plots frequency distributions for key shuttle-box metrics across individuals in the data se

Usage

plot_histograms(
  proj_data,
  bin_size_Tpref = 1,
  bin_size_Tavoid_upper = 1,
  bin_size_Tavoid_lower = 1,
  bin_size_distance = 2000,
  bin_size_shuttles = 20,
  bin_size_t_near_limits = 5,
  nr_sd = 2,
  iqr_multiplier = 1.5,
  id_col = "fileID"
)

Arguments

proj_data

Project data containg shuttle-box metrics, as output by calc_project_data

bin_size_Tpref

Bin size for temperature preference histogram, default = 1

bin_size_Tavoid_upper

Bin size for upper avoidance temperature histogram, default = 1

bin_size_Tavoid_lower

Bin size for lower avoidance temperature histogram, default = 1

bin_size_distance

Bin size for distance histogram, default = 2000

bin_size_shuttles

Bin size for number of shuttles, default = 20

bin_size_t_near_limits

Bin size for time spent near limits, default = 5

nr_sd

Number of standard deviations for outlier cutoff for normally distributed variables, default = 2

iqr_multiplier

Interquartile range multiplier for outlier cutoff for exponentially distributed variable (time near limits), default = 1.5

id_col

Column to be used for identifier labels, default = "fileID"

Value

Plot of frequency distributions for key shuttle-box metrics across individuals in dataset